requirements.txt
dash==2.5.1
gunicorn==20.1.0
tensorflow-cpu==2.9.1
Pillow==9.2.0
